Also update submodule when there is a difference
I just found out today that if a submodule is updated, then the
build job run every time the cron is called. So this started to be seen
on munin for the webbuilder of manageiq.org as a sudden spike
of CPU, and after investigation, this turned to be caused by submodules.