Make the type of builder to be a parameter of the config file
For now, as I have no idea of what is needed for jekyll or
similar, I will just try to not hardcode more than needed.

I do not want to document it since there is nothing to change
for now.

Maybe we will have to extend the script to do handle more than
bundle install, maybe more than "install" and "build" step.