Add a option to not refresh git checkout
This is used in case a user want to use Jenkins to manage
the build and so the refresh and checkout of git. While the odds
of having 1 commit between the start of the job on Jenkins
and the run of sthe script is pretty low, that can happen and
so we should make sure the script do not touch to anything.