fix YAML files to please YAMLLint
'templates/builder.yml' was renamed because it could not be excluded.
Molecule generate a list of files to pass to YAMLLint and the exclusion
list cannot be changed. As it is a syntax error due to the templating
tags, it is not possible to ask YAMLLint to ignore the error either.