Support running in container
Basically skip all the build scheduling and run the build in place.

Also allows to sync locally; the consuming service may need to be
installed in between the builder is setup and the website is build, you
may disable building and trigger it yourself later.

In case you wish to use the builder result directory directly, the
permissions have been corrected. This directory path has to be
recalculated outside this role at the moment.